- refactoring and cleaning of setting action

+++ trunk/src/gb-actions-figis/intersection/src/main/java/it/geosolutions/geobatch/figis/intersection/util/TmpDirManager.java
@@ -1,4 +1,3 @@
-
 /*
  *  GeoBatch - Open Source geospatial batch processing system
  *  http://code.google.com/p/geobatch/
@@ -24,29 +23,49 @@
 package it.geosolutions.geobatch.figis.intersection.util;
 
 import java.io.File;
+import java.io.IOException;
 
+import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TmpDirManager {
-	 private final static Logger LOGGER = LoggerFactory.getLogger(TmpDirManager.class);
-	
-	/**************
-	 * this method create the tmpdirName dir in the temporary dir
-	 * @param tmpDirName
-	 * @return
-	 */
-    public static File createTmpDir(String tmpDirName) {
-    	// creates the temporary $tmp/figis and $tmpfigis/$layername
-    	LOGGER.info("Creating the temp dir "+tmpDirName);
-    	final File sysTempDir = new File(System.getProperty("java.io.tmpdir"));
-    	String figisTmpDir = sysTempDir+System.getProperty("file.separator")+tmpDirName;
-    	File tmpDir = new File(figisTmpDir);
-    	if (!tmpDir.exists()) tmpDir.mkdir();
-    	LOGGER.info("Temp dir successfully created");
-    	return tmpDir;
+
+public class TmpDirManager
+{
+    private static final Logger LOGGER = LoggerFactory.getLogger(TmpDirManager.class);
+
+    /**************
+     * this method create the tmpdirName dir in the temporary dir
+     * @param tmpDirName
+     * @return
+     * @throws IOException
+     */
+    public static File createTmpDir(String tmpDirName) throws IOException
+    {
+        // creates the temporary $tmp/figis and $tmpfigis/$layername
+        LOGGER.info("Creating the temp dir " + tmpDirName);
+
+        final File sysTempDir = new File(System.getProperty("java.io.tmpdir"));
+        String figisTmpDir = sysTempDir + System.getProperty("file.separator") + tmpDirName;
+        File tmpDir = new File(figisTmpDir);
+        if (!tmpDir.exists())
+        {
+            tmpDir.mkdirs();
+        }
+
+        if (tmpDir.exists() && tmpDir.isDirectory() && tmpDir.canWrite())
+        {
+            LOGGER.info("Temp dir successfully created : " + tmpDir.getAbsolutePath());
+
+            return tmpDir;
+        }
+        else
+        {
+            LOGGER.error("Could not create 'figisTmpDir' (" + figisTmpDir + ")");
+            throw new IOException("Could not create 'figisTmpDir' (" + figisTmpDir + ")");
+        }
     }
-	
+
 
     /************
      *      Deletes all files and subdirectories under dir.
@@ -54,21 +73,31 @@ public class TmpDirManager {
     * If a deletion fails, the method stops attempting to delete and returns false.
      * @param dir
      * @return
+     * @throws IOException
      */
-    public static boolean deleteDir(File dir) {
-    	LOGGER.info("Deleting dir "+dir);
-        if (dir.isDirectory()) {
-            String[] children = dir.list();
-            for (int i=0; i<children.length; i++) {
+    public static boolean deleteDir(File dir) throws IOException
+    {
+
+        LOGGER.info("Deleting dir " + dir);
+        if (dir.exists() && dir.isDirectory())
+        {
+
+            /*String[] children = dir.list();
+            for (int i = 0; i < children.length; i++)
+            {
                 boolean success = deleteDir(new File(dir, children[i]));
-                if (!success) {
-                	LOGGER.error(dir+" cannot be deleted ");
+                if (!success)
+                {
+                    LOGGER.error(dir + " cannot be deleted ");
+
                     return false;
                 }
-            }
+            }*/
+            FileUtils.deleteDirectory(dir);
         }
+
         // The directory is now empty so delete it
-        return dir.delete();
-    } 
-	
+        return !dir.exists();
+    }
+
 }
